Beginning work on proper bootstrapping
