+++ bin/dev-init/init.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 export QUIET_INIT=1
 export DEBUG=0
-node -e 'require("./module/base/bin/bootstrap"); require("./bin/dev-init").init()';
+node -e 'require("./module/boot/bin/init"); require("./bin/dev-init/index").init()';
