+++ module/base/README.md
@@ -7,10 +7,29 @@ $ npm install @travetto/base
 ```
 
 Base is the foundation of all `travetto` applications.  It is intended to be a minimal application bootstrap, as well as support for commonly shared functionality. It has support for the following key areas:
+* Environmental Information
 * File Operations
 * Life-cycle Support
 * Stacktrace Management 
 * General Utilities
+
+## Environmental Information
+The framework provides basic environment information, e.g. in prod/test/dev.  This is useful for runtime decisions.  This is primarily used by the framework, but can prove useful to application developers as well. The information that is available is:
+* `prod`/`dev` - Run type flags.  These are mutually exclusive and are `boolean` flags.
+* `watch: boolean` - Does the current run support file watching.  Primarily used internally, but should be useful to indicate if the program will finish immediately or wait indefinitely.
+* `profiles: string[]` - Specific application profiles that have been activated.  This is useful for indicating different configuration or run states.
+* `debug`/`trace` - Simple logging flags.  These `boolean` flags will enable or disable logging at various levels. By default `debug` is on in `dev` or `e2e` mode, and nowhere else.  `trace` is always off by default.
+* `cwd: string` - The root of the current project, 
+* `appRoots: string[]` - The file root paths for the application, the default set is the current project. Order matters with respect to resource resolution. All paths should be relative to the project base
+* `docker` - Determine if docker support is enabled. If explicitly set, honor, otherwise it will attempt to invoke the `docker` cli and use that as it's indicator. 
+
+With respect to `process.env`, we specifically test for all uppercase, lowercase, and given case.  This allows us to test various patterns and catch flags that might be off due to casing.  That would mean that a key of `Enable_Feature` would be tested as:
+* `Enable_Feature`
+* `ENABLE_FEATURE`
+* `enable_feature`
+
+### App Information
+This basically exposes your `package.json` data as a typed data structure, useful for integrating package information into your application.
 
 ## File Operations
 The framework does a fair amount of file system scanning to auto-load files.  It also needs to have knowledge of what files are available. The framework provides a simple and performant functionality for recursively finding files. This functionality leverages regular expressions in lieu of glob pattern matching (this is to minimize overall code complexity).
