--- module/base/bin/travetto-plugin-clean.js
@@ -1,1 +1,0 @@
-require('./lib').clean();
