--- module/base/bootstrap/main.ts
@@ -1,6 +1,0 @@
-export * from './app-info';
-export * from './cache';
-export * from './env';
-export * from './fs-util';
-export * from './register';
-export * from './scan-fs';
