+++ module/base/src/app-info.ts
@@ -1,8 +1,6 @@
-// @ts-check
+import { EnvUtil, FsUtil } from '@travetto/boot';
 
-import { FsUtil } from './fs-util';
-
-const pkg = require(FsUtil.joinUnix(FsUtil.cwd, 'package.json'));
+const pkg = require(FsUtil.joinUnix(EnvUtil.cwd, 'package.json'));
 
 const subName = pkg.name.split('/').pop();
