+++ module/boot/.gitignore
@@ -1,0 +1,4 @@
+src/*.js
+src/*.d.ts
+index.d.ts
+index.js
