+++ module/boot/bin/travetto-plugin-clean.ts
@@ -1,0 +1,2 @@
+import { AppCache } from '../src/app-cache';
+AppCache.clear();
