+++ module/boot/tslint.json
@@ -1,0 +1,136 @@
+{
+  "jsRules": true,
+  "rules": {
+    "class-name": true,
+    "comment-format": [
+      true,
+      "check-space"
+    ],
+    "curly": true,
+    "forin": true,
+    "indent": [
+      true,
+      "spaces",
+      2
+    ],
+    "arrow-return-shorthand": true,
+    "import-spacing": true,
+    "new-parens": true,
+    "prefer-template": true,
+    "label-position": true,
+    "max-line-length": [
+      true,
+      180
+    ],
+    "member-access": false,
+    "member-ordering": [
+      true,
+      {
+        "order": [
+          "private-static-field",
+          "protected-static-field",
+          "public-static-field",
+          "private-static-method",
+          "protected-static-method",
+          "public-static-method",
+          "private-instance-field",
+          "protected-instance-field",
+          "public-instance-field",
+          "constructor",
+          "private-instance-method",
+          "protected-instance-method",
+          "public-instance-method"
+        ]
+      }
+    ],
+    "no-arg": true,
+    "no-bitwise": true,
+    "no-console": [
+      true,
+      "time",
+      "timeEnd"
+    ],
+    "no-construct": true,
+    "no-debugger": true,
+    "no-duplicate-variable": true,
+    "no-duplicate-imports": true,
+    "no-empty": false,
+    "no-boolean-literal-compare": true,
+    "no-this-assignment": true,
+    "no-import-side-effect": true,
+    "no-eval": true,
+    "no-for-in-array": true,
+    "no-inferrable-types": false,
+    "no-string-literal": false,
+    "no-string-throw": true,
+    "no-shadowed-variable": true,
+    "no-irregular-whitespace": true,
+    "no-invalid-this": true,
+    "no-consecutive-blank-lines": true,
+    "prefer-object-spread": true,
+    "no-invalid-template-strings": true,
+    "no-sparse-arrays": true,
+    "no-stateless-class": false,
+    "no-switch-case-fall-through": true,
+    "no-trailing-whitespace": true,
+    "no-unused-expression": false,
+    "no-use-before-declare": true,
+    "no-var-keyword": true,
+    "object-literal-shorthand": true,
+    "object-literal-key-quotes": [
+      true,
+      "as-needed"
+    ],
+    "array-type": [
+      true,
+      "array"
+    ],
+    "object-literal-sort-keys": false,
+    "one-line": [
+      true,
+      "check-open-brace",
+      "check-catch",
+      "check-else",
+      "check-whitespace"
+    ],
+    "prefer-const": true,
+    "quotemark": [
+      true,
+      "single"
+    ],
+    "radix": true,
+    "semicolon": [
+      true,
+      "always"
+    ],
+    "triple-equals": [
+      true,
+      "allow-null-check"
+    ],
+    "typedef-whitespace": [
+      true,
+      {
+        "call-signature": "nospace",
+        "index-signature": "nospace",
+        "parameter": "nospace",
+        "property-declaration": "nospace",
+        "variable-declaration": "nospace"
+      }
+    ],
+    "variable-name": [
+      true,
+      "ban-keywords",
+      "check-format",
+      "allow-pascal-case",
+      "allow-snake-case"
+    ],
+    "whitespace": [
+      true,
+      "check-branch",
+      "check-decl",
+      "check-operator",
+      "check-separator",
+      "check-type"
+    ]
+  }
+}
